engine: consolidate in-play-asset discard helper into cards (#119)
